+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
+## [1.0.0-preview-4] - 2024-06-05
 ### Changed
+- IterativeTable.Skip() was renamed to IterativeTable.IsSkipped() to avoid confusion.
+- V2FrameCapacity was renamed to V2IoCapacity to better reflect its purpose.
+- V2FragmentCapacity was renamed to V2TableCapacity to better reflect its purpose.
+- Better defaults for buffer sizes.
+
 - the `WithApplicationCertificate` on `KustoConnectionStringBuilder` was removed as it was ambiguous and not implemented correctly. Instead there are two new methods:
   - `WithAppCertificatePath` - Receives the path to the certificate file.
   - `WithAppCertificateBytes` - Receives the certificate bytes in-memory.  
@@ -15,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed Mapping Kind not working correctly with certain formats.
+- Fixed plenty of sync issues.
+- Reduced allocations.
+
 
 ## [1.0.0-preview-3] - 2024-06-05
 ### Added